A lot more of the Metadata editing functionality
