+++ resources/frontend_client/app/css/core/flex.css
@@ -20,6 +20,10 @@
     justify-content: center;
 }
 
+.justify-between {
+    justify-content: space-between;
+}
+
 .align-start {
     align-items: flex-start;
 }
