+++ resources/frontend_client/app/query_builder/popover_with_trigger.react.js
@@ -111,11 +111,9 @@ export default React.createClass({
             classes += " " + this.props.triggerClasses;
         }
         return (
-            <span>
-                <a className={classes} href="#" onClick={this.toggleModal}>
-                    {this.props.triggerElement}
-                </a>
-            </span>
+            <a className={classes} href="#" onClick={this.toggleModal}>
+                {this.props.triggerElement}
+            </a>
         );
     }
 });
